# Conflicts:
#	.idea/workspace.xml
#	src/main/java/clicknever/models/controllers/ControllerKeyBoard.java

+++ src/main/java/clicknever/models/controllers/Mouse.java
@@ -5,8 +5,8 @@ import static clicknever.models.controllers.mouse.ControllerMouse.mouseY;
 
 public class Mouse {
 
-    private float x;
-    private float y;
+    private int x;
+    private int y;
     private boolean isClick;
     private boolean isTime;
     private String name;
@@ -99,19 +99,19 @@ public class Mouse {
     }
 
 
-    public float getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public float getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(int y) {
         this.y = y;
     }
